[Move EmployeeInfo to Types.h] . Move 'EmployeeInfo'(Struct) to Types.h
